requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Enhanced Quantization
auto-gptq>=0.5.0
optimum>=1.15.0

# Performance Optimizations
flash-attn>=2.3.0
triton>=2.1.0
xformers>=0.0.22

# Data Processing
datasets>=2.15.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.15.0
mlflow>=2.8.0
rich>=13.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0
psutil>=5.9.0
GPUtil>=1.4.0

# Configuration Management
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Testing & Quality
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: For specific model architectures
einops>=0.7.0
rotary-embedding-torch>=0.3.0